python: do not raise ENOENT error on kvs.isdir()
Problem: The flux.kvs.isdir() function currently returns True if
a key is a directory.  If the key does not exist, the function raises
an exception that the key does not exist.  This is not consistent with
how Flux handles such logic.  For example, in the flux_kvsdir_isdir()
C function, false is returned if the key is missing.

If the key does not exist, return False for flux.kvs.isdir().