actions: work around sched boost warning issue
With gcc12+ the boost version we get in bookworm and jammy pops out a
maybe-uninitialized warning from one of its headers.  This seems to be
an issue with the sched build system, and should be fixed by the cmake
PR, but work around it for flux-sched check here until that's fixed
upstream.  Cross-linking to flux-framework/flux-sched#1049