testsuite: avoid use-after-free in perilog-test.so
Problem: if perilog-test.so is removed before all its internal
reactor timers have triggered, there is a potential for use-after-free.

This was encountered when making some broker changes that must have
altered timing slightly.

Before removing the plugin, wait for the clean event to be posted on
the test job, indicating that all perilog events have been processed.