libkvs: support flux_kvs_txn_is_empty()
Problem: There are times where it would be convenient to know if
a KVS transaction is empty or not.  There is no way to do that right now.

Support a new flux_kvs_txn_is_empty() function to check if the transaction
object is empty of transactions.

Add some unit tests to cover this new function.