broker: fix unprotected access to broker attrs
Problem: broker module threads copy broker attributes directly
from the broker during initialization, without locks.

Although no crashes have been observed, this is unsafe.

Note: the purpose of priming each module's attribute cache is to
avoid a synchronous RPC from flux_attr_get(3) when the broker rank
or other common attributes are fetched during module initialization.
This is not required, but is a useful optimization that should be
retained.

When the module is created, copy the broker attributes that were
cached in the broker's flux_t handle to json a object.  Then when the
module's flux_t handle is created in the module thread, copy the json
object into the module's flux_t cache.