mergify: remove status check conditions from config
Problem: The required status checks in the mergify config are
redundant with GitHub branch protections, and keeping the list of
statuses up-to-date in the mergify config is quite painful since
updates require a PR and manual merge.

Remove the status checks completely from .mergify.yml. Mergify will
still ensure that required status checks in branch protections are
satified, so this should not change any behavior.