testsuite: fix race in t0005-exec.t signal test
Problem: If the signal in the signal forwarding test in t0005-exec.t
is delivered after the "hi" sentinel is printed but before sleep(1)
is executed, then bash ignores the signal and the test will timeout
and fail.

To prevent the race use awk instead of a bash script, so that there
is no point where the delivered signal will not be handled.  The awk
script prints the sentinel then reads from a fifo in order to block
indefinitely (stdin in sharness may be closed), so there are no extra
subprocesses involved.

Fixes #5382