python: Add WatchImplementation class
Problem: The KVSWatchFuture and JobEventWatchFuture classes are
identical except for the get() and cancel() calls to get a value
or cancel the stream.

Solution: Create a new WatchImplementation class that can
act as an abstract class for both watch futures.  The futures both
need only implement their respective "get" and "cancel" implementations.

Update KVSWatchFuture and JobEventWatchFuture to be based on WatchImplementation.