flux-resource: code consistency cleanup
Problem: Usually a subcommand function is set at the end of the arguments,
but for a few subcommands it is set earlier in the command line option
parser setup.  This makes anal retentive people go cuckoo.

Solution: Setup the subcommand function at the end of the command line
option setup consistently.