flux-top: display empty title for instances that are not jobs
Problem: When the flux-top(1) title generation was updated to use the
idf58() convenience function in 2d2cf72, the check for FLUX_JOBID_ANY
was dropped, resulting in windows titled with the F58 representation
of FLUX_JOBID_ANY when flux-top(1) is connected to an instance that
is not a job. This can be confusing to the user.

Use an empty string for the title when connecting to an instance that
does not have a Flux jobid.

Fixes #5392