testsuite: ensure flux-top(1) doesn't display FLUX_JOBID_ANY title
Problem: No tests in the testsuite ensure that flux-top(1) does not
display the F58 representation of FLUX_JOBID_ANY as the window title
when connected to an instance that is not a Flux job.

Add a simple test that verifies FLUX_JOBID_ANY does not appear as
the flux-top(1) title.