matrix: add arm64 install-only builder
This is a step-1 to supporting arm64 a bit better in CI.  If this is
fast enough, maybe we can get some occasional smoke tests on top of it
or something but it should make getting dev containers going easier at
the least.

Also add qus action to register aarch64 support in binfmt, all the
emulation is already there, but it's unregistered in the runner.