job-list: refactor state transition flags
Problem: In the internal state transition struct, the flag "processed"
is a little ambiguous.  In addition, multiple locations manually
remove a state transition from the next_states list.

Solution: Split the "processed" flag into two flags, "processing" and
"finished".  The "processing" flag more clearly indicates that the state
transition is being processed.  The "finished" flag indicates the transition
is done and we can remove the struct from the next_states list.  With these
new flags, we no longer need to remove the struct from the next_states list
in so many locations.  Simply set the "finished" flag to true and remove
from the list in process_next_state().