job-manager: fix limit-duration plugin with unset duration
Problem: The limit-duration plugin explicitly allows jobs with unset
(i.e. unlimited) duration when a duration limit is set. This was
probably so that a later plugin could set a default duration, but since
duration defaults are now applied by the job frobnicator at submission
time, any default will be set by the time this plugin is called.

Reject unlimited duration jobs in the limit-duration plugin where
there is a duration limit configured.