jobtap: return error for jobspec-update of non-pending jobs
Problem: The flux_jobtap_jobspec_update_pack(3) function allows a
jobtap plugin to post a jobspec-update event to a job in any state,
but these updates are not allowed for jobs that are not pending.

Check that the current job is not in RUN or CLEANUP states, and that
the eventlog is not readonly (i.e. inactive) before scheduling a
jobspec-update event to be posted.