job-manager: reject events for jobs with readonly eventlog
Problem: The job manager sets an eventlog_readonly flag on a job for
inactive jobs and other cases where the eventlog should no longer
accept events, but this flag is not checked before queuing job events.

Check the flag before allowing an event to be queued to a job and
return EROFS (Read-only) if the flag currently disallows writes to
the eventlog.