python: allow Jobspec.setattr() key to start with attributes.
Problem: The Jobspec setattr() method always prepends 'attributes.'
the the key argument, but this can be inconvenient when the key
already contains the 'attributes.' prefix, since that prefix must
then be removed before calling jobspec.setattr().

Only prepend 'attributes.' to the key argument of setattr() if it
doesn't already contain that prefix.