job-manager: validate all job states in limit-* plugins
Problem: The limit-duration and limit-job-size plugins do not validate
jobs unless they are in the NEW state, ostensibly because job.validate
may be called after a plugin reload or job manager restart. However, it
is no longer the case that job.validate is called in these situations,
and it may be necessary to call job.validate for jobs beyond the NEW
state when processing job updates.

Drop the checks for FLUX_JOB_STATE_NEW in the limit-duration and
limit-job-size plugins.