job-manager: add job update service
Problem: There is no service in the job manager for requesting the
update of jobspec or other job parameters.

Add a new update service to the job manager. Job updates can now
be requested via a job-manager.update RPC, the payload of which
includes the target jobid and an "updates" object which follows the
jobspec-update specification in RFC 20.

Updates for a key are only allowed if a plugin callback exists for the
jobtap topic string "job.update.KEY", and the callback returns success.
If multiple keys are updated in the same request, they all must be
allowed or none will be applied.

Once updates have been validated, then the proposed modified jobspec
is sent through the job.validate plugin call stack. If the new jobspec
fails to be sucessfully validated, then the updates are rejected and
an error is returned to the requestor.

Individual plugins may request that the job.validate be skipped for a
given key by settin a 'validated' flag in the plugin OUT arguements.
However, the job.validate call will still be made if multiple keys
are being updated and not all of them set a validated flag.

Once the update is allowed and validated, then a jobspec-update event
is posted for the job and an empty success response is issued.