job-manager: add update-duration builtin plugin
Problem: Once a job is submitted the duration cannot be updated.

Add an update-duration plugin that adds a

 job.update.attributes.system.duration

callback so that jobspec duration updates are supported for pending
jobs. By default, users can update the duration of their own jobs
up to the currently configured limit, and instance owners can update
duration to any value.

The ability of the instance owner to bypass limits can be disabled
by reloading the plugin with the config parameter owner-allow-any=0.