testsuite: fix recursive jobspec-update in job-list test plugin
Problem: In the jobspec-update-job-list jobtap plugin, a test
jobspec-update event is posted from the job.state.sched callback.
However, in the future jobspec-update events will cause a job in SCHED
state to transition back to PRIORITY, and when the job then goes back
to SCHED state, the jobspec-update event will be emitted again, the
job will transition back to PRIORITY again, creating an infinite loop.

Ensure the test plugin only emits a jobspec-update once during the
test to avoid infinite recursion. This will require the plugin to be
reloaded to continue working after one job, but at this point that
is not necessary.