job-manager: call job.update plugin stack after jobspec-update
Problem: There is no way for a jobtap plugin to get notified of a
jobspec update after the jobspec updates have been applied. Jobs only
transition back to PRIOITY state from SCHED, so the job.state.priority
callback will not always be sufficient, and subscribing directly to
the jobspec-update event would require the plugin to manually apply
updates, and may not capture other ways a jobspec or job might be
updated in the future.

Introduce a 'job.update' callback topic which is called after
any jobspec update has been applied. If the job is transitioning
back to the PRIORITY state, this callback will be called before
the job.state.priority topic so that plugins may adjust internal
state that would normally be established prior to the first call
to job.state.priority.