job-manager: deny updates for jobs that bypass update validation
Problem: When the instance owner updates a guest job in order to bypass
validation (e.g. to update duration of a job beyond current limits),
a future job update of a different attribute may fail because the job
will be revalidated. This causes a confusing error that is unrelated
to the user's request.

When a job update bypasses validation, the update request is made by
the instance owner, and the job user is not the instance owner, mark
the job as immutable to prevent future updates by the job owner. This
not only results in a less confusing error

 "job is immutable due to previous instance owner update"

and also prevents the need to track which attribute updates have
bypassed validation in past updates, which could be complex and could
introduce unintended consequences.