job-list: do not cache jobspec_tasks
Problem: The "tasks" object of a jobspec is currently cached in struct job.
In the future, the tasks object could be changed via a jobspec-update event,
so we should not cache it.  In addition, the tasks object is only used
in one location, and only under certain circumstances.  So it is not necessary
to read it on every parse of the jobspec.

Solution: Remove cache of "tasks" object.  Only parse it when necessary.