job-list: splice out load of the jobspec
Problem: In the near future jobspec-update events will be allowed
to update any field of the jobspec.  Current code is not very reusable
for supporting this.

Solution: Splice out the load of the initial jobspec into its own
function.  Now parse_jobspec() only parses the internally loaded
jobspec, allowing it to be re-used by other parts of the code.