job-list: re-initialize jobspec fields
Problem: In the near future jobspec-update events will be allowed to
update any field in the jobspec.  This means that certain fields
may be parsed differently or not at all on another parse.  We need to
re-initialize some fields so that subsequent parses behave as intended.

Re-initialize a number of fields as necessary if they affect parsing or
if they are not parsed when parsing the jobspec again.