cmd: flux-perilog-run: send signals to entire process tree
Problem: The flux-perilog-run script delivers signals only to the
pid of the direct child process when processes time out, but if
the subprocess is a shell running some foreground tasks, the signal
may be ignored, or the subprocess may be killed but its children
remain running.

Instead of just killing the direct child on timeouts, attempt to
send the signal to the process and all its descendants.