testsuite: cover non attributes jobspec updates
Problem: The current coverage of "jobspec-update" events in
t2260-job-list.t only covers fields under the "attributes"
dictionary.  We should cover more than that.

Update the jobspec-update-job-list to update the command to
run for the job instead of the job name.  The end result will be
that the job name viewed from job-list will be changed.