cmd: flux-perilog-run: skip offline nodes in per-rank mode
Problem: If ranks are offline then a prolog/epilog executed on
those ranks in --exec-per-rank mode fails and the node is drained
with a generic drain reason "prolog/epilog for job X failed" which
is unhelpful.

Skip offline ranks in --exec-per-rank mode, i.e. do not attempt
to run a prolog or epilog on nodes that are currently down. Log a
message when this occurs so there's some hint in the logs, and exit
with nonzero status (since prolog/epilog on that node was not run).

Fixes #5391