cmd: flux-perilog-run: add -D, --drain-offline option
Problem: Ranks that are offline in --exec-per-rank mode are skipped,
but some sites may want to drain these skipped ranks with a specific
error message.

Add a -D, --drain-offline option to the flux-perilog-run epilog and
prolog subcommands which causes a drain message of

 offline for JOBID [prolog|epilog]

to be set on any ranks skipped because they are offline.