flux-exec: add --imp-run option
Problem: There is no easy way to use `flux-imp run` with flux-exec(1).

Add a new --with-imp option to flux-exec(1) which simply prepends
"/path/to/flux-imp", "run" to the command to execute.

Set a use_imp global boolean when flux-imp is being used, either
explicitly via the --with-imp option, or if argv0 has a basename of
'flux-imp'. This will be useful in future work.