flux-perilog-run: add --with-imp option
Problem: To run the per-rank prolog or epilog under flux-imp run so
that they execute with privileges requires explicitly adding the full
path to the flux-imp on the flux-perilog-run command line in config.

Add a --with-imp option to both the prolog and epilog subcommands of
flux-perilog-run and pass the option along to flux-exec(1) if used
so that the commands are executed under `flux-imp run`.