libpmi: add ops->warn server callback
Problem: there is no simple way for the PMI server to
issue a shell_warn() when it could be helpful to users.

Add a warn() operation callback that can be populated in
the shell.

Call warn() on rank 0 if a client attempts to connect with
PMI version 2.  Hint that the application may be using slurm's
libpmi2.so.