libpmi: use stdio in simple_client
Problem: the libpmi "simple client" reads single characters
from the server socket in order to build a line without the
need to buffer characters.

Although this approach is functional, it is a best practice to
minimize the number of read(2) calls using buffering when possible.

Use libc buffered I/O instead.