flux-job: get current jobspec representation
Problem: Internally within flux, the jobspec can be updated through
jobspec-update events.  When the user runs "flux job info" to get
the jobspec, it is getting the jobspec stored in the KVS and it may
not represent what is actually used.  In addition, the jobspec returned
from "flux job info" will be inconsistent to the information returned
from "flux jobs".

Solution: When retrieving the jobspec, also retrieve the job eventlog.
Internally update the jobspec through "jobspec-update" events and output
the final result to give the user a representation of what the jobspec
looks like within flux.

Add a --base option to perform the old behavior, which is to just get
the jobspec stored in the KVS.

Fixes #5411