jobtap: add jobtap_register_builtin()
Problem: There is no way for job manager components to register an
internally defined jobtap builtin plugin, but this could be useful
because it allows a dynamically loadable plugin access to the job
manager component internals.

Add a new type of jobtap builtin defined by a jobtap_builtin_ex
struct along with a new builtins_ex list to supplement the existing
builtin plugins.

The difference between this type of plugin and other jobtap builtins is
that it allows an opaque argument to be passed to the initialization
callback which can be propagated to registered jobtap plugin
callbacks, giving the plugin the ability to access internal component
data without needing to expose these internals via new flux_jobtap*
functions.

The new jobtap_register_builtin() function can be called from within
any job manager component to register and load a new builtin plugin.
These external builtins are then searched and loaded in addition to
the "internal" builtins.