job-manager: perilog: improve prolog exception errors
Problem: The exception note generated by the perilog job-manager
plugin when a job manager prolog fails is not easily interpreted
since it doesn't decode the raw wait status of the failed command.
Since this note is displayed to the user, it would be better to try
to print a more helpful message.

Decode the wait status of the prolog command and attempt to detect
if it was signaled or exited with a nonzero exit code. If the prolog
was signaled by a SIGTERM, then add a parenthetical that the prolog
may have timed out or the job was canceled.