flux-job: suppress status line after fatal job exception
Problem: When a pending job gets a fatal job exception, the exception
is printed and then followed by the attach statusline, e.g.:

 $ flux run hostname
 3.336s: job.exception type=prolog severity=0
         prolog killed by signal 15 (timeout or job canceled)
 flux-job: ƒZoj78siD5 waiting for job prolog                   00:00:03

This is a bit messy and could cause confusion.

Suppress the status line display after a fatal job exception. The
job is no longer waiting on anything except to exit, so the status
is not helpful.