python: refactor kvslookup jobspec flags
Problem: The job kvslookup code uses a number of flags to manage
special case "jobspec" lookups.  While this is fine when there is
one single lookup special case, future special cases make the
addition of more flags very cumbersome.

Refactor the logic to remove flags.  Instead, maintain a copy of
the inputted lookup keys and a copy of the keys used for the actual
lookup.  Use this information instead of the multiple created flags
to determine special case handling.