top: use streaming job-stats RPC
Problem: flux-top(1) resends a job-list.job-stats RPC every
heartbeat, but now a streaming option is available.

Use a streaming RPC so only one request needs to be sent.

Move the job-stats RPC handler to top.c to prepare for using
it to drive other queries, in place of job state transition event
messages.