flux-job: get current R representation
Problem: Internally within flux, R can be updated via resource-update
events.  But when the user runs "flux job info" to get R, it is getting
the version stored in the KVS and may not be representative of the
"viewed" R.  The result may not be consistent other tools such as "flux jobs".

Solution: When retrieving R, also retrieve the job eventlog.  Update R
through "resource-update" events and output that as the final result.
Get the stored version of R only if the --base option is specified.

This change breaks several tests in t2232-job-info-security.t that
created a dummy value of R.  Instead of reading a dummy value of R,
create a "dummy" key and read that.

Fixes #5425