python: support updated lookups of R
Problem: The "flux job info" command will present an updated
view of "R" when it is looked up, or a "base" view when --base
is specified.  This is supported in the python job/kvslookup.py
module with jobspec, but not with R.

Support updated and base lookups of R as well as jobspec.