job-manager: update: refactor update_job()
Problem: update_job() might be used in the future to post an update
after some asynchronous work is completed, but it doesn't have the
right args to do this, nor does it respond to the user after the
update event is posted (or fails to post).

Refactor the job_update() function so it posts the update event
and sends the response to the user, allowing it to be used from any
context to complete a job update.