job-manager: update: support feasibility check for updates
Problem: For some job attributes it may be required to determine if
the update is feasible in addition to allowed and valid, but there
is currently no way for an update to trigger the feasibility check.

Since jobtap plugin callbacks can't easily do asynchronous work, and
to allow feasibility checks on a set of updates as a whole instead of
individual updates for one key in a set, add a flag in the job.update.*
plugin out arguments that allows an update plugin to require that
the update pass feasibility before being posted to the job eventlog.

If any plugin sets `feasibility` nonzero in the plugin out arguments,
then a feasibility RPC is issued for the fully updated jobspec. Since
the feasibility request is an RPC, a new container for the request
and other update data is created so that the actual update and reply
to the original request can be handled after the RPC returns.