job-manager: update: allow plugins to amend updates
Problem: Some job update plugins may need to extend or amend the set
of updates applied to a job. For example, a queue update may require
a change to the current job constraints.

Allow plugins to modify the updates that will be applied to a job by
setting an 'updates' key in the plugin out arguments. These updates
are collected from all plugins, then applied to the updates with
json_object_update() before validation and feasibility are performed.