job-manager: queue: store queue `requires` config in jobq struct
Problem: Configured queue constraints are not saved along with
other job queue data in the job manager jobq struct. However, this
information will be useful in the future when validating updates of
a job's queue.

Add a requires field to `struct jobq` to save the job queue requires
array for later use.