job-manager: jobtap: make update failure errnos consistent
Problem: The job update related jobtap function jobtap_job_update()
and jobtap_validate_updates() set an error string that is eventually
returned in the error response to the initiator of the request,
but they do not set a specific errno, which can result in unexpected
errno values from flux_rpc_get(3), potentially causing confusion.

Explicitly set errno to EINVAL in these functions when the update
is either not supported or is rejected by a plugin, or is determined
to be invalid by one or more `job.validate` callbacks. This provides
a consistent errno for these cases in case a caller is checking.

Update one test in t2290-job-update.t which demonstrated a confusing
errno: EPERM was returned for an update of a duration to a negative
value (EINVAL is much more appropriate in this case).