content: make the content cache a broker module
Problem: the content service embedded in the broker is
resource-intensive under load, potentially impacting broker
performance.  Furthermore, it is a complex, optional service
rather than an integral broker component.

Move the content service to a broker module where it is less
tightly coupled to the broker and has a private reactor thread.

For easier review, this commit does not make substantive changes
to the code.  Changes are limited to:
- move source code
- change header guards to reflect new location
- comment out internal broker attribute code that would not compile
- add main.c containing module boiler plate
- update build system

Subsequent commits will convert from the internal attribute interface
to the Flux API interface.

Load/unload content in the installed rc1/rc3 scripts.