job-manager: correct fsd output in error message
Problem: An error message in the limit-duration plugin uses the
output from fsd_format_duration_ex() with a precision of 1.  The
precision of 1 leads to the typical output to be in exponent form.
For example, "30m" is output as "3e+01m".

Solution: Increase the precision from 1 to 2, which should make the
typical output look fine.

Fixes #5414