copy content.hash value
Problem: the result of flux_attr_get ("content.hash") is
stored without copying in several places, but this assumes
that the attribute is immutable.  If the attribute is not,
then the value is clobbered on the next call to flux_attr_get().

Since the content service that owns "content.hash" is moving to
a broker module, it will not be able to set the IMMUTABLE flag on
that attribute.

Update users to make a copy of the value.

Add a 'hash_type' parameter to content_store_get_blobref() instead
of fetching it internally (since now every fetch would be a
synchronous RPC).  Update users.