testsuite: increase wait-event timeout
Problem: A recent run of t2611-debug-emulate.t had the test

"debugger: SIGCONT can unlock a job in stop-tasks-in-exec"

fail due to a presumed timeout.

Solution: Increase the test timeout from 10 seconds to 100 seconds.

Update one test that needs to a test to timeout in order to pass, and
we don't want to wait for 100 seconds.  Hard code that timeout to 2 seconds.