flux-keygen: drop libsodium requirement
Problem: libsodium is required by flux-keygen but it is not checked
in configure.ac.

All we do with libsodium at this point is add its version to certificate
metadata.  This was just "provenance", never actually used for anything,
and not worth the extra dependency.  Rather than fix configure.ac, drop
the libsodium version from flux-keygen(1) certificates.

Fixes #5445