rc: use posix shell in shebang
Problem: if BASH_ENV is set to LLNL's /etc/bashrc, flux fails to
start, exiting at the beginning of rc1 without any clue as to why.

Example output:

  rc1.0: /etc/flux/rc1 Exited (rc=2) 1.0s
  rc3.0: flux: Error committing finish event: No service matching kvs.commit is registered
  rc3.0: flux-content: content.flush: Function not implemented
  rc3.0: flux-module: remove content-sqlite: No such file or directory
  rc3.0: /etc/flux/rc3 Exited (rc=1) 2.0s

The problem results from our use of the "#!/bin/bash -e" shebang
and applies to any BASH_ENV script that is not "errexit clean".

As explained by Py Watson:

  I think that errexit is not supposed to apply to the BASH_ENV processing
  even when turned on via the bash command line option. However, a
  particular builtin Bash command causes the errexit processing to become
  active if turn on. It just some happens that one of the profile.d
  scripts uses that command. Once active, the next unchecked failure of a
  command leads to bash exiting.

  The /etc/profile.d/flatpak.sh script has this line:
    if command -v flatpak > /dev/null; then

Solution: switch the shebang to "#!/bin/sh -e".  The POSIX shell does not
source anything in non-interactive mode.

Because we were using the bash "nullglob" setting, a function needed to
be added to get proper glob processing in the rc subdirs.

Fixes #5447