etc: fix rc errors with multiple rcX.d files
Problem: When multiple rc files are present in the rc1.d directory
rc1 fails because the glob is not split due to a prior setting of
IFS=:. This causes the error:

 rc1.0: /etc/flux/rc1: 107: /etc/flux/rc1.d/01-sched-fluxion
 rc1.0: /etc/flux/rc1.d/02-cron: not found

Since IFS is only applied because presumably the result of the glob
is passed back from the listfiles() function, remove the use of the
listfiles() function and use the glob as the loop argument directly.

Handle the nomatch case by skipping files that do not exist. This
should work because glob will only expand to files that exist or the
glob itself (which will not exist).

Fixes #5452