job-info: support new update-lookup/watch service
Problem: In the future, several services will need to know a job's
resources and know the updates that would apply to them.  This would
currently require users to read R, read the eventlog, and then apply
`resource-update` events to R.  Some other users would also need to
know when there are changes to R, necessitating watching the eventlog
for future resource-update changes.

It would be nice if a service did this as there will be multiple users.

Solution: Support a new job-info.update-lookup service and
job-info.update-watch streaming service.  It currently supports only
the key "R", but can be extended to other keys in the future.

The job-info.update-lookup service will read R and the eventlog for
a job.  it then apples all resource-update changes to R and returns
the result.

job-info.update-watch service will do the same as the above, but if
the job is not completed, it will continue to watch the eventlog for
future resource-update events.  On each new resource-update event,
a new R will be streamed back to the caller.  This continues until
the job ends or the caller cancels the stream.

Fixes #5451