broker: get a zeromq context directly from libzmq
Problem: Now that all zeromq sockets in the broker are allocated
with zmq_socket() instead of zsock_new(), zsys is no longer
invoked internally by czmq to track unclosed sockets, and thus
does not need to be initialized.

Get the zeromq context from zmq_ctx_new() instead of zsys_init().

Instead of zsys_shutdown() (via atexit(3)), call zmq_ctx_term().
However, don't call it if any modules had to be canceled, as it
blocks when 0MQ sockets are leaked.

The broker now exits with a return code of 1 when sockets are leaked.