libzmqutil: don't call zsock_resolve()
Problem: now that all 0MQ sockets are created with
zmq_socket() rather than zsock_new(), there is no need
for libzmqutil functions to be "polymorphic".

Drop the calls to zsock_resolve() to extract the raw socket
from the void * socket argument when a zsock_t is passed in.

In classes where that was the last CZMQ call, drop the czmq.h
include directive.