configure: use distutils.version if available
Problem: The Python packaging module is not standard, but configure
uses it to compare versions, leading to an implicit new dependency.

Since distutils is not removed until Python 3.12, try distutils first
then fall back to the packaging package. This will allow configure
to continue to work on existing systems while also not breaking on
Python 3.12.