librlist: support FLUX_HWLOC_XMLFILE to read hwloc from file
Problem: Loading topology with hwloc_topology_load(3) on systems
with many cores can be very slow -- even moreso when many processes
are trying to load the topology due to sysfs locking. Ironically,
this can make the testsuite more likely to fail on systems with many
cores than on systems with fewer cores when the testsuite is run in
parallel, since we not only have multiple brokers being launched in
parallel for each tests, but many tests run in parallel, resulting
in possibly hundereds of topology loads occuring simultaneously.

While an environment variable to force hwloc to load topology
from an XML file exists (HWLOC_XMLFILE), this variable has
proven unreliable. Therefore, introduce a Flux equivalent:
FLUX_HWLOC_XMLFILE. When this environment variable is set to file
containing hwloc XML, all hwloc_topology_load(3) from the system will
be skipped, instead using the contents of the file.  This can then
be used to workaround the issue with slow topology.