librlist: ensure system info is added to hwloc loaded from file
Problem: The librlist code currently assumes that HostName is set on
the Machine object of a loaded hwloc topology, but this information
is not required by hwloc so it may be missing when topology has been
loaded from an XML file. For example, the HostName and other Machine
attributes are not set when creating a fake topology with lstopo(1).

If HostName is not found in the topology after loading from an XML
file, add it using the result of uname(2). Add other system info
provided by uname(2) as well for completeness.