testsuite: add missing pipe between echo and rpc
Problem: In t2231-job-info-eventlog-watch.t, there is a missing
pipe between an echo of a json payload and the 'rpc' helper tool,
so a payload was not actually being sent.  The test was just echoing
the name of the command.

Add a pipe between the echo and 'rpc' helper tool to initiate the
correct payload and test.