testsuite: test invalid input with streaming flag
Problem: A test in t2231-job-info-eventlog-watch.t does not set the
streaming flag in an invalid input test.  This test happens to work
because the code's "is the input valid" check is done before the
"is the streaming flag set" test.  But for full correctness, the
streaming flag should be set.

Solution: Use the 'rpc_stream' tool in the test, which sets the
streaming flag.