doc: group man3 pages by topic
Problem: the man3 section is very long.

Add a level to the toctree and group entries into sections.
Add new index_*.rst to Makefile.am EXTRA_DIST.
Add index.rst to Makefile.am EXTRA_DIST (it was missing).

Suppressing the include fragments in 'common' from the table
of contents seems tricky now that they are referenced from
multiple indices, so drop the header titles from the fragments
and add them to the documents that include them.