python: improve begin-time dependency representation
Problem: When begin-time dependencies are displayed in flux-jobs
via the dependencies JobInfo property, they are simply rendered as
a timestamp in seconds, which is not very user-friendly.

Add a DependencyList class which can post-process the obtained list of
dependencies and improve the representation of select types. For now,
just convert `begin-time` dependencies to human readable format. If a
begin time occurs on the current day, just display the time, otherwise
display the date and time.

Fixes #5465